feat: add session restoration with CLI flags and in-app picker
Add --continue and --sessions CLI flags for resuming previous Claude
sessions. Includes interactive SessionPicker component, session index
reader, /sessions builtin command, and AppPhase state management.
