+++ source/commands/__tests__/executor.test.ts
@@ -26,6 +26,7 @@ function makeContext(
 			addMessage: vi.fn(),
 			exit: vi.fn(),
 			clearScreen: vi.fn(),
+			showSessions: vi.fn(),
 			sessionStats: {
 				metrics: {
 					modelName: null,
